requirements.txt
torch>=2.1.0
transformers>=4.46.0,<5.0.0
trl>=0.12.0
peft>=0.7.0
datasets>=2.14.0
anthropic>=0.18.0
accelerate>=0.24.0
evaluate>=0.4.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
tqdm>=4.66.0
pyyaml>=6.0
scikit-learn>=1.3.0
